+++ lib/src/features/main/main_screen_route.dart
@@ -1,5 +1,6 @@
 import 'package:flutter/material.dart';
 import 'package:provider/provider.dart';
+import 'package:where_to_go_today/src/features/auth/code/code_screen.dart';
 
 import 'main_screen.dart';
 import 'main_screen_vm.dart';
@@ -8,13 +9,15 @@ import 'main_screen_vm.dart';
 class MainScreenRoute extends MaterialPage<void> {
   MainScreenRoute()
       : super(
-          child: Provider<MainScreenVm>(
-            create: MainScreenVm.new,
-            child: Builder(
-              builder: (context) => MainScreen(
-                vm: context.read<MainScreenVm>(),
-              ),
-            ),
-          ),
+          child: const CodeScreen(),
+
+          //  Provider<MainScreenVm>(
+          //   create: MainScreenVm.new,
+          //   child: Builder(
+          //     builder: (context) => MainScreen(
+          //       vm: context.read<MainScreenVm>(),
+          //     ),
+          //   ),
+          // ),
         );
 }